pyproject.toml
[project]
name = "asset-placer-eval-standalone"
version = "0.1.0"
description = "Standalone JSON generation and eval-loop workspace (Isaac/Omni-free)"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "openai>=1.0.0",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
  "opencv-python-headless==4.10.0.84",
  "matplotlib>=3.8.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
package = false
